Conflicts:
	Diabetik/Diabetik-Info.plist

+++ Diabetik/View Controllers/Events/UAInputBaseViewController.m
@@ -557,7 +557,7 @@
     if(image)
     {
         NSTimeInterval timestamp = [NSDate timeIntervalSinceReferenceDate];
-        NSString *filename = [NSString stringWithFormat:@"%d", (NSInteger)timestamp];
+        NSString *filename = [NSString stringWithFormat:@"%ld", (long)timestamp];
         
         __weak typeof(self) weakSelf = self;
         [[UAMediaController sharedInstance] saveImage:image withFilename:filename success:^{
